requirements.txt
torch
torchvision
albumentations
tqdm
pytorch-lightning
Pillow
numpy
pandas
trimesh
vedo
opencv-python